PP-10004 M1 support for integration tests
- moved static method to utility class, this fixes a unit test failure in CI environments where no docker engine was available
- removed Docker Client dependency, this is no longer needed as TestContainers instruments all ephemeral test dependencies
